fix: re-add DisplayUnit to A320
